started implementing depth for Kinect driver

+++ src/main/java/edu/mines/acmX/exhibit/input_services/hardware/devicedata/DepthImageInterface.java
@@ -1,7 +1,8 @@
 package edu.mines.acmX.exhibit.input_services.hardware.devicedata;
 
 import java.nio.ByteBuffer;
+import java.nio.ShortBuffer;
 
 public interface DepthImageInterface extends DeviceDataInterface {
-	public ByteBuffer getDepthImageData();
+	public ShortBuffer getDepthImageData();
 }
